refactor: rename files and routes to kebab-case
