+++ src/components/ui/string-collapsible.tsx
